fix(ObjC): wrong package name in commons
