+++ commons/src/main/java/fr/insideapp/sonarqube/apple/commons/SourceLinesProvider.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package fr.insideapp.sonaqube.apple.commons;
+package fr.insideapp.sonarqube.apple.commons;
 
 import org.sonar.api.internal.apachecommons.io.input.BOMInputStream;
 import org.sonar.api.utils.log.Logger;
