+++ commons/src/main/java/fr/insideapp/sonarqube/apple/commons/antlr/CustomTreeVisitor.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package fr.insideapp.sonaqube.apple.commons.antlr;
+package fr.insideapp.sonarqube.apple.commons.antlr;
 
 import org.antlr.v4.runtime.tree.AbstractParseTreeVisitor;
 import org.antlr.v4.runtime.tree.ParseTree;
