+++ commons/src/main/java/fr/insideapp/sonarqube/apple/commons/antlr/ParseTreeItemVisitor.java
@@ -15,42 +15,13 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package fr.insideapp.sonaqube.apple.commons;
+package fr.insideapp.sonarqube.apple.commons.antlr;
 
-public final class SourceLine {
+import org.antlr.v4.runtime.tree.ParseTree;
+import org.sonar.api.batch.sensor.SensorContext;
 
-    private final int count;
-    private final int start;
-    private final int end;
-    private final int line;
+public interface ParseTreeItemVisitor {
+    void apply(ParseTree tree);
 
-    public SourceLine(final int line, final int count, final int start, final int end) {
-        this.line = line;
-        this.count = count;
-        this.start = start;
-        this.end = end;
-
-    }
-
-    @Override
-    public String toString() {
-        return "SourceLine [line=" + line + ", count=" + count + ", start=" + start + ", end=" + end + "]";
-    }
-
-    public int getLine() {
-        return line;
-    }
-
-    public int getCount() {
-        return count;
-    }
-
-    public int getEnd() {
-        return end;
-    }
-
-    public int getStart() {
-        return start;
-    }
-
+    void fillContext(SensorContext context, AntlrContext antlrContext);
 }
