+++ commons/src/main/java/fr/insideapp/sonarqube/apple/commons/issues/RepositoryRuleParser.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package fr.insideapp.sonaqube.apple.commons.issues;
+package fr.insideapp.sonarqube.apple.commons.issues;
 
 import org.sonarsource.analyzer.commons.internal.json.simple.JSONArray;
 import org.sonarsource.analyzer.commons.internal.json.simple.JSONObject;
