+++ commons/src/main/java/fr/insideapp/sonarqube/apple/commons/tests/AppleTestsSensor.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package fr.insideapp.sonaqube.apple.commons.tests;
+package fr.insideapp.sonarqube.apple.commons.tests;
 
 import org.sonar.api.batch.sensor.Sensor;
 import org.sonar.api.batch.sensor.SensorContext;
