+++ commons/src/main/java/fr/insideapp/sonarqube/apple/commons/tests/StaxParser.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package fr.insideapp.sonaqube.apple.commons.tests;
+package fr.insideapp.sonarqube.apple.commons.tests;
 
 import com.ctc.wstx.api.WstxInputProperties;
 import com.ctc.wstx.stax.WstxInputFactory;
