+++ swift-lang/src/main/java/fr/insideapp/sonarqube/swift/lang/antlr/SwiftAntlrContext.java
@@ -17,9 +17,9 @@
  */
 package fr.insideapp.sonarqube.swift.lang.antlr;
 
-import fr.insideapp.sonaqube.apple.commons.SourceLine;
-import fr.insideapp.sonaqube.apple.commons.SourceLinesProvider;
-import fr.insideapp.sonaqube.apple.commons.antlr.AntlrContext;
+import fr.insideapp.sonarqube.apple.commons.SourceLine;
+import fr.insideapp.sonarqube.apple.commons.SourceLinesProvider;
+import fr.insideapp.sonarqube.apple.commons.antlr.AntlrContext;
 import fr.insideapp.sonarqube.swift.lang.antlr.generated.Swift5Lexer;
 import fr.insideapp.sonarqube.swift.lang.antlr.generated.Swift5Parser;
 import org.antlr.v4.runtime.CharStream;
