+++ swift-lang/src/test/java/fr/insideapp/sonarqube/swift/lang/issues/swiftlint/SwiftLintReportParserTest.java
@@ -17,8 +17,7 @@
  */
 package fr.insideapp.sonarqube.swift.lang.issues.swiftlint;
 
-import fr.insideapp.sonaqube.apple.commons.issues.ReportIssue;
-import fr.insideapp.sonarqube.swift.lang.issues.swiftlint.SwiftLintReportParser;
+import fr.insideapp.sonarqube.apple.commons.issues.ReportIssue;
 import org.junit.Test;
 
 import java.util.List;
